+++ src/components/indices/IndexBulk.vue
@@ -31,6 +31,25 @@
         @done="emitAndCloseMenu('reload')"
       />
 
+      <row-menu-action
+        method="indexPutSettings"
+        :method-params="{ indices: selectedIndices, body: { index: { blocks: { write: true } } } }"
+        :text="t('indices.index_row.options.set_readonly.text')"
+        :growl="t('indices.index_row.options.set_readonly.growl', { index: selectedIndices.join(', ') })"
+        :confirm="t('indices.index_row.options.set_readonly.confirm', { index: selectedIndices.join(', ') })"
+        icon="lock"
+        @done="emitAndCloseMenu('reload')"
+      />
+
+      <row-menu-action
+        method="indexPutSettings"
+        :method-params="{ indices: selectedIndices, body: { index: { blocks: { write: false } } } }"
+        :text="t('indices.index_row.options.set_writable.text')"
+        :growl="t('indices.index_row.options.set_writable.growl', { index: selectedIndices.join(', ') })"
+        icon="lock_open"
+        @done="emitAndCloseMenu('reload')"
+      />
+
       <q-separator />
 
       <row-menu-action